requirements.txt
amqp==5.1.1
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickhouse-connect==0.6.2
kombu==5.3.0
lz4==4.3.2
prompt-toolkit==3.0.38
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
zstandard==0.21.0
